fix: an org can only delete its own pets at delete-pet use case

+++ src/use-cases/delete-pet.ts
@@ -1,9 +1,11 @@
 import { PetsRepository } from '@/repositories/pets-repository'
 import { Pet } from '@prisma/client'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { UnauthorizedError } from './errors/unauthorized.error'
 
 interface DeletePetUseCaseRequest {
   id: string
+  orgId: string
 }
 
 interface DeletePetUseCaseResponse {
@@ -15,11 +17,16 @@ export class DeletePetUseCase {
 
   async execute({
     id,
+    orgId,
   }: DeletePetUseCaseRequest): Promise<DeletePetUseCaseResponse> {
     const pet = await this.petsRepository.findById(id)
 
     if (!pet) {
       throw new ResourceNotFoundError()
+    }
+
+    if (pet.org_id !== orgId) {
+      throw new UnauthorizedError()
     }
 
     await this.petsRepository.delete(id)
